css/css-logical/logicalprops-inline-size.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Logical Properties: {max-,min-}inline-size</title>
<link rel="author" title="Xu Xing" href="mailto:openxu@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-logical-props-1/#logical-dimension-properties">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<style>
#tests div {
  border: 1px solid #000;
}
#div1 {
  inline-size: 40px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}
#div2 {
  inline-size: 100px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}
#div3 {
  inline-size: 120px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}

#tests p {
  border: 1px solid #000;
}
#p1 {
  inline-size: 100px;
  width: 50px;
}
#p2 {
  width: 50px;
  inline-size: 100px;
}
#ref_p1 {
  width: 50px;
}
#ref_p2 {
  width: 100px;
}

.table {
  border: 1px solid #000;
  display: table;
}
.tablecell {
  display: table-cell;
}
#table1_cell {
  inline-size: 40px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: red;
}
#table2_cell {
  inline-size: 100px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: blue;
}
#table3_cell {
  inline-size: 120px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: green;
}
</style>

<div id="tests">
  <div id="div1" data-expected-client-width="50" data-expected-client-height="0"></div>
  <div id="div2" data-expected-client-width="100" data-expected-client-height="0"></div>
  <div id="div3" data-expected-client-width="100" data-expected-client-height="0"></div>

  <p id="p1" data-expected-client-width="50" data-expected-client-height="0"></p>
  <p id="p2" data-expected-client-width="100" data-expected-client-height="0"></p>

  <div class="table">
    <div class="tablecell" id="table1_cell" data-expected-client-width="50" data-expected-client-height="100"></div>
  </div>
  <div class="table">
    <div class="tablecell" id="table2_cell" data-expected-client-width="100" data-expected-client-height="100"></div>
  </div>
  <div class="table">
    <div class="tablecell" id="table3_cell" data-expected-client-width="100" data-expected-client-height="100"></div>
  </div>
</div>

<script>
// Check that inline-size < min-inline-size or min-inline-size < inline-size <= max-inline-size or inline-size > max-inline-size
checkLayout("#div1, #div2, #div3");

// Check that width override inline-size and vice versa
checkLayout("#p1, #p2");

// Check that inline-size < min-inline-size or min-inline-size < inline-size <= max-inline-size or inline-size > max-inline-size in table
checkLayout(".tablecell");
</script>
